Adds smtp max retries mechanism
If the sent_try count for a message exceeds the smtp_max_retries
threshold, the event is logged to file and marked for deletion so that
no more delivery attempts are made.

Further, if the smtp_max_retries config is set, GLPI will begin logging
a warning when an SMTP delivery fails.