Add missing MSIN field to simcards
MSIN is the last 8 or 10 digits of IMSI
IMSI is the combination of:
- MCC (Mobile Country Code) (3 digits)
- MNC (Mobile Network Code) (2 or 3 digits)
- MSIN (Mobile Subscriber Identification Number) (8 or 10 digits)

Having 3 digits length MCC and 3 digits length MNC and 10 digits length MSIN is not allowed
because the full length is at most 15 digits.

Possible combinations are:
3 digits MCC + 2 digits MNC + 10 igits MSIN = 15 digits
3 digits MCC + 3 ditigs MNC + 8 digits MSIN = 14 digits
3 digits MCC + 2 digits MNC + 8 digits MSIN = 13 digits

MCC and MNC are stored in the line operator itemtype
MSIN is stored in a items_simcards itemtype. To save the full IMSI, one must
create a simcard component, instanciate it, fill its MSIN field and associate it to a line operator
with MCC and MNC fields populated.

add missing upgrade code
add tooltip for MSIN field
add label tag to fields of components
remove specific js
ensure  is define in more cases
fix duplicated search options in line itemtype