Bug fix for when dropdown not found
Handle case when the class is not found and $tmpname is not an array.

How to reproduce: Autoloaders for plugins, such as GenericObject, are not available for some ajax requests, such as /ajax/comments.php.

Bug fix when class not found

How to reproduce: Autoloaders for plugins, such as GenericObject, are not available for some ajax requests, such as /ajax/comments.php.

Closes #4635