Update crontask.class.php
When loading Plugins into tasks, the name of the class is considered in the default "capitalize". So if I have a plugin in the "test" folder, when trying to execute the Cron task it tries to access the "Test" folder and the following error is displayed:
"Can not redeclare plugin_teste_check_config () ...."

To solve, I added the method "strtolower" in loading Plugins.
Att. Maykon Douglas