Add Go To card action to open full form
Opens the full item form as it previously did when clicking the card title.
While the current behavior of opening the main form in a dialog when clicking the card title may be enough for a lot of cases, it is nice to have the option to go to the full form.