Try using current location as map default
Try to make it easier to assign geolocation data by defaulting the map view to the current location (Or as close as we can approximate) if the user allows us to use their location. If not, fall back on the previous defaults.
If we can get geolocation data, the map zoom level is adjusted based on the reported accuracy.
This only works over HTTPS or localhost. If those requirements don't apply to the user, they will see the original functionality.
Tested on a desktop (very-low accuracy) and a cell phone (high accuracy) and the zoom levels seemed appropriate.