Update Ticket.php
When a new ticket is created, a notification is sent to actors. Its type is "New Ticket". If a rule closes (rejects) the ticket upon creation, this "New Ticket" notification template is still used. In a perfect world, there should be both New Ticket and Close Ticket notifications sent, but as actually this is not possible, Close Ticket notification seems to be more accurate.