Changing notification type for new tickets directly closed upon creation (#12111)
* Update Ticket.php

When a new ticket is created, a notification is sent to actors. Its type is "New Ticket". If a rule closes (rejects) the ticket upon creation, this "New Ticket" notification template is still used. In a perfect world, there should be both New Ticket and Close Ticket notifications sent, but as actually this is not possible, Close Ticket notification seems to be more accurate.

* Update Ticket.php

* Update src/Ticket.php

OK for me.

Co-authored-by: Cédric Anne <cedric.anne@gmail.com>

* Wrong indent on 2121?

* Update src/Ticket.php

* Mutualize code with Problem and Change

Co-authored-by: Cédric Anne <cedric.anne@gmail.com>