fix(reservations): Don't compare reservations date range endpoints with equality
When the reservation free period search tool is looking for a free spot
it fails to find a spot adjacent to an existing reservations.

It's because it uses equality when comparing existing reservations' date
range endpoints. So it can, for example, reject a (free) starting range
of 16:00:00 if there is an existing reservation that ends at 16:00:00.

This commit removes the equality comparison by changing ">=" operator to
">" and "<=" operator to "<".

Issue: #GH-13548