feat(reservations): Adds a location-based filter on reservation search tool
This commits adds a "Location" filter on the "Find a free item in a
specific period" reservation tool.

It adds, to the existing search form, a location dropdown that is
pre-filled with currently connected user's location (if any) and filter
found reservation items to those at (or under) provided location.

Issue: #UE-2149