Update Item_Ticket.php
In the "params'" array we calculate the entities_id field, but when this function is called, its options arguments have also "entities_id". We need to take care not to overwrite with this value to avoid wrong entity restriction to user active entity and ancestors instead of tickets entity and ancestors.