Update NotificationTargetCommonITILObject.php
When a followup is retrieved by an mail collector, existing session does not set "canView" to yes. So  itil obejcts' documents are not added to the notification template data. On the other hand, as the same function is used during itil followups creations from the user interface, rights checks should not be bypassed all the time...