Notifications triggered by email followups don't mention tickets documents added after the initial ticket creation (#15454)
* Update NotificationTargetCommonITILObject.php

When a followup is retrieved by an mail collector, existing session does not set "canView" to yes. So  itil obejcts' documents are not added to the notification template data. On the other hand, as the same function is used during itil followups creations from the user interface, rights checks should not be bypassed all the time...

* Update src/NotificationTargetCommonITILObject.php

Co-authored-by: Cédric Anne <cedric.anne@gmail.com>

---------

Co-authored-by: Cédric Anne <cedric.anne@gmail.com>