Adding 'Processors number' in the computers overview list
In the computers list, computer.php, there is no possibility of showing the number of CPUs. If CPUs have the same name (as in nearly all computers) this is not visible anyhow in the list (while number of cores and threads can be displayed already). This would be important for various tasks and invoicing.

The translation of 'Processors number' is already present in GLPI, used for VMs.