Extended check for valid image
The local check is only intended for developing purposes if a plugin is not (yet) published on the marketplace. Otherwise the icon defined in the xml file of the plugin will be used as symbol.

Note: Most plugins in the marketplace uses png files. The plugin "Webhook" uses a svg file instead.